pyproject.toml
[project]
name = "secureagent"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "openai (>=1.78.1,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "autogen (>=0.9.1.post0,<0.10.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "weaviate-client (>=4.14.3,<5.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
]

[tool.poetry]
packages = [{include = "secureagent", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
